made a better error message when calling classroom that does not exist
